Move link stuff to shared package.
