+++ shared/links-metadata/index.ts
@@ -1,4 +1,7 @@
 export { SynctexStore } from "./src/SynctexStore";
 export { TexStore } from "./src/TexStore";
 export { migrateBlock, migrateSelection } from "./src/migrate";
+export { type Link, parseLink } from "./src/link";
+
 export * from "./src/types";
+export * from "./src/metadata";
