+++ src/utils/blockIdsEqual.ts
@@ -1,4 +1,4 @@
-import type { ISynctexBlockId } from "@fluffylabs/types";
+import type { ISynctexBlockId } from "@fluffylabs/links-metadata";
 
 export function blockIdsEqual(blockId1: ISynctexBlockId, blockId2: ISynctexBlockId): boolean {
   return blockId1.pageNumber === blockId2.pageNumber && blockId1.index === blockId2.index;
